move static visit us content into it’s own file
