+++ common/views/components/LandingBody/LandingBody.js
@@ -1,40 +1,18 @@
 // @flow
 import { type Weight } from '../../../services/prismic/parsers';
 import { defaultSerializer } from '../../../services/prismic/html-serializers';
-import { classNames, font, grid } from '@weco/common/utils/classnames';
-import { getTodaysVenueHours } from '@weco/common/services/prismic/opening-times';
-import SpacingComponent from '../SpacingComponent/SpacingComponent';
+import { classNames } from '@weco/common/utils/classnames';
 import FeaturedText from '../FeaturedText/FeaturedText';
 import Layout12 from '../Layout12/Layout12';
 import Layout8 from '../Layout8/Layout8';
-import Icon from '@weco/common/views/components/Icon/Icon';
 import Space from '@weco/common/views/components/styled/Space';
-import OpeningTimesContext from '@weco/common/views/components/OpeningTimesContext/OpeningTimesContext';
-import FindUs from '@weco/common/views/components/FindUs/FindUs';
 import SpacingSection from '../SpacingSection/SpacingSection';
 import GridFactory from '../GridFactory/GridFactory';
 import WobblyEdge from '../WobblyEdge/WobblyEdge';
 import SectionHeader from '../SectionHeader/SectionHeader';
 import Card from '../Card/Card';
 import FeaturedCard from '../FeaturedCard/FeaturedCard';
-
-type ContainerProps = {|
-  children: any,
-|};
-
-const Container = ({ children }: ContainerProps) => (
-  <SpacingSection>
-    <SpacingComponent>
-      <div
-        className={classNames({
-          'body-part': true,
-        })}
-      >
-        <Layout8>{children}</Layout8>
-      </div>
-    </SpacingComponent>
-  </SpacingSection>
-);
+import VisitUsStaticContent from './VisitUsStaticContent';
 
 type BodySlice = {|
   type: string,
@@ -86,95 +64,8 @@ const Body = ({ body, isDropCapped, pageId }: Props) => {
         'basic-body': true,
       })}
     >
-      {/*  'Visit us' specific content */}
-      {pageId === 'XvxzjhQAAJmq1t__' && ( // TODO move this out
-        <OpeningTimesContext.Consumer>
-          {openingTimes => (
-            <Container>
-              <div className="grid">
-                <div
-                  className={classNames({
-                    [grid({ s: 12, l: 6, xl: 6 })]: true,
-                    [font('hnl', 4)]: true,
-                  })}
-                >
-                  <FindUs />
-                </div>
-                <div
-                  className={classNames({
-                    [grid({ s: 12, l: 6, xl: 6 })]: true,
-                    [font('hnl', 4)]: true,
-                  })}
-                >
-                  <div className="flex">
-                    <Space
-                      as="span"
-                      h={{ size: 'm', properties: ['margin-right'] }}
-                    >
-                      <Icon name="clock" extraClasses={`float-l`} />
-                    </Space>
-                    <div
-                      className={classNames({
-                        [font('hnl', 5)]: true,
-                        'float-l': true,
-                      })}
-                    >
-                      <h2
-                        className={classNames({
-                          [font('hnm', 5)]: true,
-                          'no-margin': true,
-                        })}
-                      >{`Today's opening times`}</h2>
-                      <ul className="plain-list no-padding no-margin">
-                        {openingTimes.collectionOpeningTimes.placesOpeningHours.map(
-                          venue => {
-                            const todaysHours = getTodaysVenueHours(venue);
-                            return (
-                              todaysHours && (
-                                <Space
-                                  v={{
-                                    size: 's',
-                                    properties: ['margin-top'],
-                                  }}
-                                  as="li"
-                                  key={venue.name}
-                                >
-                                  {venue.name.toLowerCase() === 'restaurant'
-                                    ? 'Kitchen '
-                                    : `${venue.name} `}
-                                  {todaysHours.opens ? (
-                                    <>
-                                      <time>{todaysHours.opens}</time>
-                                      {'—'}
-                                      <time>{todaysHours.closes}</time>
-                                    </>
-                                  ) : (
-                                    'closed'
-                                  )}
-                                </Space>
-                              )
-                            );
-                          }
-                        )}
-                      </ul>
-                      <Space
-                        v={{
-                          size: 's',
-                          properties: ['margin-top'],
-                        }}
-                        className={`no-margin`}
-                      >
-                        <a href="/opening-times">Opening times</a>
-                      </Space>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </Container>
-          )}
-        </OpeningTimesContext.Consumer>
-      )}
-      {/*  end of 'Visit us' specific content */}
+      {pageId === 'XvxzjhQAAJmq1t__' && <VisitUsStaticContent />}
+
       {featuredText && (
         <Layout8>
           <SpacingSection>
@@ -187,6 +78,7 @@ const Body = ({ body, isDropCapped, pageId }: Props) => {
           </SpacingSection>
         </Layout8>
       )}
+
       {sections.map((section, index) => {
         const sectionTheme = sectionThemes[index % sectionThemes.length];
         const hasFeatured =
